略作修改 无大改动

+++ src/main/java/com/example/coin/controller/RelationshipController.java
@@ -26,7 +26,6 @@ public class RelationshipController {
     private RedisUtil redisUtil;
 
     //错误信息
-    private static final String ENTITY_EXIST = "该关系节点已存在";
     private static final String ID_NOT_EXIST = "该关系节点ID不存在";
 
     @RequestMapping(path = "/getRelationship", method = RequestMethod.GET)
@@ -45,20 +44,7 @@ public class RelationshipController {
     public ResponseVO addRelById(@RequestParam(value = "fromId")String fromId,
                                  @RequestParam(value = "toId")String toId,
                                  @RequestParam(value = "name")String name){
-        Entity entity1, entity2;
-        try{
-            entity1 = (Entity) redisUtil.get(ENTITY_REDIS_PREFIX+fromId);
-            if(entity1 == null) entity1 = entityService.findEntityById(fromId);
-            entity2 = (Entity) redisUtil.get(ENTITY_REDIS_PREFIX+toId);
-            if(entity2 == null) entity2 = entityService.findEntityById(toId);
-            if(entity1 == null || entity2 == null)throw new Exception();
-        }catch (Exception exception){
-            return ResponseVO.buildFailure(ID_NOT_EXIST);
-        }
-        redisUtil.set(ENTITY_REDIS_PREFIX+fromId, entity1);
-        redisUtil.expire(ENTITY_REDIS_PREFIX+fromId, TWO_HOURS_IN_SECOND);
-        redisUtil.set(ENTITY_REDIS_PREFIX+toId, entity2);
-        redisUtil.expire(ENTITY_REDIS_PREFIX+toId, TWO_HOURS_IN_SECOND);
+        if (someEntityIsNull(fromId, toId)) return ResponseVO.buildFailure(ID_NOT_EXIST);
         relationship newRel = relationshipService.addRelationship(fromId, toId, name);
         redisUtil.set(RELATIONSHIP_REDIS_PREFIX+fromId+"-"+toId, newRel);
         redisUtil.expire(RELATIONSHIP_REDIS_PREFIX+fromId+"-"+toId, TWO_HOURS_IN_SECOND);
@@ -67,20 +53,7 @@ public class RelationshipController {
 
     @RequestMapping(path = "/delRelationship", method = RequestMethod.POST)
     public ResponseVO deleteRelById(@RequestParam(value = "fromId")String fromId, @RequestParam(value = "toId")String toId){
-        Entity entity1, entity2;
-        try{
-            entity1 = (Entity) redisUtil.get(ENTITY_REDIS_PREFIX+fromId);
-            if(entity1 == null) entity1 = entityService.findEntityById(fromId);
-            entity2 = (Entity) redisUtil.get(ENTITY_REDIS_PREFIX+toId);
-            if(entity2 == null) entity2 = entityService.findEntityById(toId);
-            if(entity1 == null || entity2 == null)throw new Exception();
-        }catch (Exception exception){
-            return ResponseVO.buildFailure(ID_NOT_EXIST);
-        }
-        redisUtil.set(ENTITY_REDIS_PREFIX+fromId, entity1);
-        redisUtil.expire(ENTITY_REDIS_PREFIX+fromId, TWO_HOURS_IN_SECOND);
-        redisUtil.set(ENTITY_REDIS_PREFIX+toId, entity2);
-        redisUtil.expire(ENTITY_REDIS_PREFIX+toId, TWO_HOURS_IN_SECOND);
+        if (someEntityIsNull(fromId, toId)) return ResponseVO.buildFailure(ID_NOT_EXIST);
         redisUtil.del(RELATIONSHIP_REDIS_PREFIX+fromId+"-"+toId);
         relationshipService.deleteRelationById(fromId, toId);
         return ResponseVO.buildSuccess();
@@ -110,5 +83,23 @@ public class RelationshipController {
         for(relationship each:allRels){
             relationshipService.addRelationship(each);
         }
+    }
+    //extract addRelById和deleteRelById中的重复代码成方法
+    private boolean someEntityIsNull(@RequestParam("fromId") String fromId, @RequestParam("toId") String toId) {
+        Entity entity1, entity2;
+        try{
+            entity1 = (Entity) redisUtil.get(ENTITY_REDIS_PREFIX+fromId);
+            if(entity1 == null) entity1 = entityService.findEntityById(fromId);
+            entity2 = (Entity) redisUtil.get(ENTITY_REDIS_PREFIX+toId);
+            if(entity2 == null) entity2 = entityService.findEntityById(toId);
+            if(entity1 == null || entity2 == null)throw new Exception();
+        }catch (Exception exception){
+            return true;
+        }
+        redisUtil.set(ENTITY_REDIS_PREFIX+fromId, entity1);
+        redisUtil.expire(ENTITY_REDIS_PREFIX+fromId, TWO_HOURS_IN_SECOND);
+        redisUtil.set(ENTITY_REDIS_PREFIX+toId, entity2);
+        redisUtil.expire(ENTITY_REDIS_PREFIX+toId, TWO_HOURS_IN_SECOND);
+        return false;
     }
 }
